Update README.md
Description

This model has been developed for the purpose of identifying quotes from two famous speakers, Menachem Begin and Yitzhak Rabin. The model is based on a pretrained language model.

Usage

This model can be used for personal or professional purposes to check whether a given sentence is attributed to either Menachem Begin or Yitzhak Rabin.